Add workaround to pass tests based on current functionality.
Obviously is most cases, this is *not* what you want to do with a test.
However, the cause of the issue is in upstream packages (specifically,
the DatasetSchema class in ogd-common), and their interaction with
outdated naming conventions of some of the files on the server.

For that reason, putting in the temporary workaround until that issue
with naming/parsing names is resolved, since it's not the FileAPI itself
causing the mismatch between data and what we *want* the data to be.